refactor: move the trace's name sanitizing out of exec/tap.ts
Deriving the reportable command and flag names had accumulated three lookup
tables and two helpers in the dispatch module, which imported `renderOptionsFor`
and `TapCommandOptionSchema` for no other reason.

Move them to `tap/reported-invocation.ts`, whose single export states the
invariant it exists to hold: nothing a user typed reaches the wire. `exec/tap.ts`
is left with dispatch, and `beginTapTrace` now takes the reported invocation
rather than a loose command and flag list.

The tables cannot live in events.ts itself — `commands/*` imports `output.ts`,
which imports `events.ts`, so reading `tapCliCommands` from there would resolve
undefined at module init. events.ts takes only the type.
